gradle-update
fixed some project errors
