--- .gitmodules
@@ -1,4 +1,0 @@
-[submodule "opensdk"]
-	path = opensdk
-	url = https://github.com/WaxMoon/opensdk.git
-	branch = main
