+++ app/.gitignore
@@ -1,56 +1,1 @@
-.externalNativeBuild
-import-summary.txt
-
-*.class
-*.dex
-.sync/
-
-*.iws
-*.ipr
-*.iml
-target/
-.idea/
-.gradle/
-.disk/
-release/
-build/
-spoon/
-
-#mac temp file
-.idea/
-__MACOSX
-.DS_Store
-._.DS_Store
-
-#for eclipse
-.settings/
-local.properties
-*gen/
-*.classpath
-*/bin/
-.project
-
-#temp file
-*.bak
-
-*.pmd
-sh.exe.stackdump
-
-.vs/
-.vscode/
-
-*.log
-*.ctxt
-.mtj.tmp/
-
-hs_err_pid*
-
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
-*.cxx
-*.cfg
-
+/build
