+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.waxmoon"
-    android:sharedUserId="org.waxmoon.github">
+    xmlns:tools="http://schemas.android.com/tools">
 
     <!--高危权限-->
     <uses-permission android:name="android.permission.MANAGE_EXTERNAL_STORAGE" />
@@ -118,23 +117,23 @@
         android:multiArch="true"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
-        android:label="${label}"
+        android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.GithubMultiApp" >
+        android:theme="@style/Theme.GithubMultiApp">
         <activity
             android:name=".AppListActivity"
             android:exported="true"
-            android:theme="@style/Theme.GithubMultiApp.NoActionBar" >
+            android:theme="@style/Theme.GithubMultiApp.NoActionBar"
+            >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
         <provider
-            android:name="org.waxmoon.MoonProvider"
+            android:name="com.alex.pro.MoonProvider"
             android:authorities="${applicationId}:org.waxmoon.MoonProvider"
             android:exported="false"
             android:grantUriPermissions="true">
