+++ app/src/main/java/com/alex/pro/MoonProvider.java
@@ -1,0 +1,8 @@
+package com.alex.pro;
+
+import androidx.core.content.FileProvider;
+
+public class MoonProvider extends FileProvider
+{
+
+}
