--- app/src/main/java/org/waxmoon/MoonProvider.java
@@ -1,6 +1,0 @@
-package org.waxmoon;
-
-import androidx.core.content.FileProvider;
-
-public class MoonProvider extends FileProvider {
-}
