+++ assist/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.waxmoon"
+
     android:sharedUserId="org.waxmoon.github">
 
     <!--高危权限-->
@@ -117,7 +117,8 @@
         android:supportsRtl="true"
         android:usesCleartextTraffic="true">
         <meta-data android:name="metadata.master.pkg" android:value="${masterPkg}"/>
-        <activity android:name=".MainActivity">
+        <activity android:name=".MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
